move `migrate-arbitrary-value-to-bare-value` to core
